Node selector for SFC CR node reconciler
We need the node selector to tell the operator where to read the SFC CR
and to deploy the network function pod. Current the design is that the
daemon has a split brain and it needs the node selector to make sure
that one daemon doesn't unintendedly deploy on the wrong node.

Current Limitations!: Because the SFC pod being rendered has the same
name and namespace. Only 1 network function pod can be created cluster
wide. This limitation is not the effect of this change.

Signed-off-by: William Zhao <wizhao@redhat.com>