Allow SFC CR to create pod in default namespace
We monitor for pods in the default and operator namespace. This allows
us to manage the network function pods we create by the SFC CR.

We continue to monitor for all other resources in the dpu-operator
namespace. This is the default behaviour of most operators.

This commit also fixes the namespace replacement in role binding files.

Signed-off-by: William Zhao <wizhao@redhat.com>