Compare each field in the DPU CR's status field for changes
Prior to this commit, the DPU CR when using deep equal on status
would flood logs with updates. However these updates contain no actual
changes. The reason is due to using deep equal on a data structure
containing metadata from k8s. In this case it would be LastTransitionTime.

Instead we need to compare each relevent field in the Condition under
the status field. This stops the flooding of updates and the logs for
the update. We make sure for the "Ready" type that the condition is up
to date or created. Currently "Ready" is the only supported condition.

Signed-off-by: William Zhao <wizhao@redhat.com>