Move more targets into taskfile
We're already tracking dependencies in the taskfile, and we are unifying
everything so that we can fully rely on it instead of partially relying
on make and partially on task. Automation still uses make, and those
targets then call into task. Rename a few targets while at it, and move
checks to check.yaml

Signed-off-by: Balazs Nemeth <bnemeth@redhat.com>