marvell: add workaround to disable tx/rx hardware offload on veth pairs
As we have veth interfaces, there is no "hardware" to offload to.
Instead, we attach those interfaces to OVS. The veth pair then
sends incomplete packets without a checksum, and OVS doesn't
fill them in either. The result is a broken checksum and broken TCP.

It is not clear why this happens and why this is a problem. Shouldn't
OVS detect that the packet received via the veth port has the checksum
undefined? So that it can complete it before sending out out via the SDP
hardware interface (or offload the calculation of the checksum to the
SDP interface)? That does not seem to happen, and that might be a bug in
the SDP driver (MDC-123).

As workaround, avoid this problem by disabling hardware offloading for
"tx" and "rx".

https://issues.redhat.com/browse/MDC-122
https://issues.redhat.com/browse/MDC-123
See-also: https://arthurchiao.art/blog/ovs-deep-dive-5-datapath-tx-offloading/