Add ready handler
By ensuring that we run at least one detection cycle, any DPU CRs
that should be synced will be synced before we move the the DPU
Operator Config CR to ready state. That way, there is not sleep
needed in any automation that sets up the operator and waits for
the DPUs to be available between waiting for the operator config
to move to ready state and waiting for the first DPU CR to be
available. We also call the daemon "not ready" as long as it's not
really up (i.e. it's running the detection cycles).

The main intent is to remove a tricky place where race conditions
could occur with anything that interacts with the DPU Operator.

Signed-off-by: Balazs Nemeth <bnemeth@redhat.com>