tools: more strictly select lines for incremental Dockerfile
It seems this "incremental.go" would be nicer as 10 lines shell script.
At least the pattern matching would be nicer with sed and a pattern that
more carefully matches the lines we care about.

Anyway. But it seems wrong trying to mangle all lines that contain
"COPY" and "--from=builder", if they don't also contain a
"/workspace/bin".

After all, when I have a builder stage in the Dockerfile, I will want to
leave artifacts inside /workspace. In the later stage, I want to do:

  COPY --from=builder /workspace/my-artifact .

Previously, the incremental tool would match that line and replace it
with

  COPY /workspace/my-artifact .

That is wrong.

Now, we can leave stuff in /workspace/ (not /workspace/bin/) and that
will be taken from the builder stage.

Note that currently there are no Dockerfiles for which this change would
make a difference. But soon there will be and we will need a way to not
wrongly include certain lines.