dockerfile: support building "Dockerfile.mrvlVSP.rhel" outside of go-task
The way "Dockerfile.mrvlVSP.rhel" is build is not good. The go-task
"build-cpagent-service" would have generated "bin/cp-agent.service".
Later, the "Dockerfile.mrvlVSP.rhel" would take it from there.

But this means that the "Dockerfile.mrvlVSP.rhel" cannot be build
directly. It needs this additonal task to generates build artifacts
outside the container build first. This cannot work with ART, which
takes the Dockerfile, mangles it, and builds it directly.

Hack around that. Now, the go-task "prepare-cpagent-service" generates a
file "bin/cp-agent-image.txt" that only contains the image URL. The
"Dockerfile.mrvlVSP.rhel" will take the image URL from there and use it
to generate the "cp-agent.service". It will always do so, even if the
"bin/cp-agent-image.txt" file is missing. With this change, a plain
`buildah build -f Dockerfile.mrvlVSP.rhel` will at least pass.

Of course, when we build the Dockerfile without running the go-task,
then the image URL will not be valid. So while this allows to build the
container image, it's a broken image. But this current approach of
injecting the image URL is fundamentally broken and can't work for
images in Operatorhub. We need to inject the image URL via the manifest
and enviornment variable. The proper fix for that will come in
https://issues.redhat.com/browse/MDC-107.

This change merely allows to build the container, it is still broken.