Wait for VFs when trying to reference their PCIe Address/Name
The reason is that VFs when created may need time from the driver to
properly setup sysfs amongst other kernel data structures. Therefore it
is possible to have created the sriov_numvfs, with the VFs being in a
transitory state towards proper creation.

We introduce some wait functions that should be used when VFs are first
created/referenced.

Signed-off-by: William Zhao <wizhao@redhat.com>