manifest: uncomment image-references to use ART images
The ART build tooling uses this file to patch the manifest
file. See the script at [1].

We failed to properly onboard various images in ART. This is now in the
process of being fixed with [2].

Uncomment the image-references to make use of the ART images that are
now being built.

[1] https://github.com/openshift-eng/art-tools/blob/99cda969e7fec2e34b35b1cd393dc6c72f3c0ae4/doozer/doozerlib/backend/rebaser.py#L2065
[2] https://issues.redhat.com/browse/NHE-1571