Fix Issue where CmdDel occurs on a VF that was not plumbed on the DPU
CmdDel can be called when CmdAdd fails; however if CmdAdd failed and the
netconf is not cached then the subsequent CmdDel would not be able to
find the non-existant cache. This is ok because the CmdAdd would have
been able to properly cleanup before returning.

The issue is that if CmdDel continues to call DeleteBridgePort, the DPU
would get an nonsense request. We shouldn't propagate DeleteBridgePort
when CreateBridgePort would have never executed due to CmdAdd failing.

Signed-off-by: William Zhao <wizhao@redhat.com>