Fix timing issue with waiting for operator deployment
There is an error message that occasionally pops up:
   "error: no matching resources found"
when waiting for pods in the openshift-dpu-operator namespace.

This happens because the openshift-dpu-operator namespace exists,
but the operator's controller (which is a Deployment resource) is
still in the process of creating the ReplicaSet and, subsequently,
the actual pods.

The original oc wait command looks for pods and finds none, it
correctly reports "error: no matching resources found" and exits
with fail.

Signed-off-by: William Zhao <wizhao@redhat.com>