Support 2 DPUs from same vendor in single cluster deployment
Since the VSP name takes the DPU Platform Identity which is a fixed
string per DPU model, that means no more 1 DPU per model can be deployed
in single cluster. Fix this by making the Intel PTL function take the
MAC address of a known interface and actually honor the DPU Identifier
concept (code commentary) that it should be unique identification of the
DPU platform. Other DPUs, need to find that unique identification and
implement the DPU Identifier function.

Signed-off-by: William Zhao <wizhao@redhat.com>