marvell: set interface down after removing SDP interface from OVS bridge
This is in response to RHEL-108203, where passing traffic on SDP
interfaces (while the VF on host side doesn't exist) might cause the
problem.

We already only set interfaces up when attaching them to the bridge. At
that point, the host clearly created the VFs already. Now, when
detaching the interface from the bridge, set it down again. This is to
avoid a potential sutuation where the interface might be up while the
host having no VF.

https://issues.redhat.com/browse/RHEL-108203