tft: Update to latest version of k8s traffic flow tests
Specially we want to pull in the changes that fixes a race condition:
https://github.com/ovn-kubernetes/kubernetes-traffic-flow-tests/pull/231

We should make sure that the IperfServer is listening on the tcp/udp socket
first before allowing the IperfClient to connect.

This makes sure that we don't get into any race conditions with a slow start
up of iperf3 -s.

There were changes to how the k8s traffic flow test works and in order
to move to the latest k8s traffic flow test tool, we need to update the
pod jinja contents.

Signed-off-by: William Zhao <wizhao@redhat.com>